requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
openai
google-genai
chromadb
pypdf
pymupdf
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-chroma
streamlit